move is local admin helper function into role helper file
