--- src/app/shared/helpers/user-helper.ts
@@ -1,17 +1,0 @@
-import { OrganizationRight } from '../models/organization-right.model';
-
-export function rightsIncludesLocalAdminInOrg(
-  organizationRights: OrganizationRight[] | undefined,
-  organizationUuid: string | undefined
-): boolean {
-  if (!organizationUuid || !organizationUuid) {
-    return false;
-  }
-  const localAdminEnumValue = 1;
-  return (
-    organizationRights
-      ?.filter((right) => right.organizationUuid === organizationUuid)
-      .map((right) => right.role)
-      .includes(localAdminEnumValue) ?? false
-  );
-}
